#2015020902 登録時の処理を追加

+++ classes/VizualizerTwitter/Module/Account/SaveOperators.php
@@ -36,10 +36,13 @@ class VizualizerTwitter_Module_Account_SaveOperators extends Vizualizer_Plugin_M
         $post = Vizualizer::request();
         if ($post["add"] || $post["save"]) {
             $loader = new Vizualizer_Plugin("Twitter");
+            $companyIds = array();
             $operatorIds = array();
             if(is_array($post["operator_id"])){
                 foreach($post["operator_id"] as $operatorId){
-                    if($operatorId > 0){
+                    if(substr($operatorId, 0, 1) == "*" && substr($operatorId, 1) > 0){
+                        $companyIds[] = substr($operatorId, 1);
+                    }elseif($operatorId > 0){
                         $operatorIds[] = $operatorId;
                     }
                 }
@@ -54,7 +57,7 @@ class VizualizerTwitter_Module_Account_SaveOperators extends Vizualizer_Plugin_M
 
                 // 登録から外れたグループを削除
                 foreach($operators as $operator){
-                    if(!in_array($operator->operator_id, $operatorIds)){
+                    if(!in_array($operator->operator_id, $operatorIds) && !in_array($operator->company_id, $companyIds)){
                         $operator->delete();
                     }
                 }
